+++ src/providers.rs
@@ -1,8 +1,6 @@
 use anyhow::{anyhow, bail, Context, Result};
 use clap::Parser;
-use fiberplane::provider_runtime::spec::types::{
-    Blob, LegacyProviderRequest, LegacyProviderResponse, ProviderConfig, ProviderRequest,
-};
+use fiberplane::provider_runtime::spec::types::{Blob, ProviderConfig, ProviderRequest};
 
 #[derive(Parser)]
 pub struct Arguments {
@@ -13,61 +11,15 @@ pub struct Arguments {
 pub async fn handle_command(args: Arguments) -> Result<()> {
     use SubCommand::*;
     match args.sub_command {
-        Invoke(args) => handle_invoke_command(args).await,
         Invoke2(args) => handle_invoke2_command(args).await,
     }
 }
 
 #[derive(Parser)]
 pub enum SubCommand {
-    /// Invoke a provider with the legacy provider protocol
-    Invoke(InvokeArguments),
-
     /// Invoke a provider with the new provider protocol
+    #[clap(alias = "invoke")]
     Invoke2(Invoke2Arguments),
-}
-
-#[derive(Parser, Debug)]
-pub struct InvokeArguments {
-    /// Path to the provider WASM file
-    #[clap(long, short)]
-    pub provider_path: String,
-
-    /// JSON encoded request that will be sent to the provider
-    #[clap(long, short)]
-    pub request: String,
-
-    /// JSON encoded config that will be sent to the provider
-    #[clap(long, short)]
-    pub config: String,
-}
-
-async fn handle_invoke_command(args: InvokeArguments) -> Result<()> {
-    let request: LegacyProviderRequest =
-        serde_json::from_str(&args.request).context("unable to deserialize request")?;
-    let config = parse_config(&args.config).context("unable to deserialize config")?;
-
-    let wasm_module = std::fs::read(args.provider_path)
-        .map_err(|e| anyhow!("unable to read wasm module: {:?}", e))?;
-
-    let runtime = fiberplane::provider_runtime::spec::Runtime::new(wasm_module)
-        .map_err(|e| anyhow!("unable to create runtime: {:?}", e))?;
-
-    let result = runtime.invoke(request, config).await;
-
-    match result {
-        Ok(LegacyProviderResponse::Error { error: err }) => {
-            Err(anyhow!("Provider failed: {:?}", err))
-        }
-        Ok(val) => match serde_json::to_string_pretty(&val) {
-            Ok(val) => {
-                println!("{val}");
-                Ok(())
-            }
-            Err(e) => Err(anyhow!("unable to serialize result: {:?}", e)),
-        },
-        Err(e) => Err(anyhow!("unable to invoke provider: {:?}", e)),
-    }
 }
 
 #[derive(Parser, Debug)]