+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format of this file is based on [Keep a Changelog](https://keepachangelog.co
 
 ## [Unreleased]
 
+### Removed
+
+- Removed support for the legacy provider protocol. The `provider invoke`
+  command has now become an alias for the new `invoke2` subcommand.
+
 ## [2.6.0]
 
 ### Added